Major changes, added the interrupts module
Added the interrupts module, with the function for handling the device's interrupts.

Wrote a draft for the device interrupts dispatcher and for the timer interrupt service routine. Also
wrote a procedure to update the interval time.

Wrote a main module, which exports all the global kernel variables.

Wrote a types header, which exports all the types used in the kernel.

Added cpu time fields to the pcb type.

Wrote a function to increase the priority of all processes in a pcbQueue.

Wrote a syscall header to export the ystem calls code to the rest of the kernel.

Added the new modules to the makefile.