Some architectural fixes and whatnot
A new field has been added to the pcb to act as a semaphore to wait on a child process. The former
method for waiting a child process would cause a dangling reference.

The dispatch method now takes the address of a state to save in the pcb of the running process, in
case this needs to be swapped (for example when the timeslice expires).

Two new methods have been added. One is called restoreRunningProcess and is called when the
interrupt handling routine has done its job and needs to restore the former running process. The
other is called insertInReady and handles the insertion of a process in the ready queue in a
preemptive manner (the running process may be sent back to the ready queue if its priority is not
high enough).

The V operation has been modified to take a state_t* argument, for the same reason the dispatch
method was modified.

Restructured both the Interrupt handler and the Syscall handler: now the entry point is the
beginning of the funciton and the only exit point is the end of the function (unlike previously
where a P operation could be the exit point of the handler).

The timer handler has been modified to reflect prevoius changes and to return a next-to-be-loaded
process to better handle the timeslice policy of YAEOS.

Added some type checks to the handleTimer function. Futhermore, if the last aging/pseudclock
deadline has been passed (negative interval) the timer is not set, so that the interrupt is
generated once more and handled appropriately.

The created processes would not be correctly linked according to their gerarchic relationships
(parent/child). This has been fixed.

The increasePriority method would not preempt the running process if a new process, with a higher
priority, would be yielded by the priority increase. This has been fixed.