Fixed a huge bug that blocked me for a week
Fixed a huge bug that blocked me for a week, i forgot to update the timer before going into a
waiting state to wait for an interrupt to start the machine again when there are no processes left.
This has been fixed.

Fixed some accounting bugs with the softBlocked and activePcbs counters.

Fixed the removeProcQ function, it would not set the removed pcbs' next pointer to NULL (this would
fuck the ready queue up).