Bugfixes on kernel and test code
Added a function to the scheduler module to handle the passup. This way all the operations of the
passup are handled inside the scheduler module (There are no mode naked LDST).

Modified the sysHandler, tlbHandler and pgmTrapHandler to use the passup function.

In p2test the p2 function used an old specification of the SYS6 (the one frome JaeOS). Fixed.

In the kernelTimeAccounting function the billed process was the running one, instead of the one
given as an argument.

Added a check on SYS6 to see if any argument is NULL.