Bugfixes and comments
Added libdiv.o to linking for the divisions.

Moved some accounting to the end of the interrupt Handler.

Bugfix in aging handling, it would not increase the priority of all processes once but many times
and the reinsertion of the runningPcb wasn't right.

Bugfixes in the list module.

Small modifications in checkForDeadlock.